Refactoring

+++ PersonalFinancialTracker/src/main/java/dev/personal/financial/tracker/repository/admin/AdminRepositoryImpl.java
@@ -1,5 +1,7 @@
 package dev.personal.financial.tracker.repository.admin;
 
+import dev.personal.financial.tracker.exception.user.UserAlreadyBlockedException;
+import dev.personal.financial.tracker.exception.user.UserNotFoundException;
 import dev.personal.financial.tracker.model.User;
 import dev.personal.financial.tracker.repository.user.UserRepository;
 
@@ -7,6 +9,10 @@ import lombok.RequiredArgsConstructor;
 
 import java.util.List;
 
+/**
+ * Реализация интерфейса {@link AdminRepository}.
+ * Обрабатывает административные функции, такие как блокировка и удаление пользователей.
+ */
 @RequiredArgsConstructor
 public class AdminRepositoryImpl implements AdminRepository {
     private final UserRepository userRepository;
@@ -19,14 +25,21 @@ public class AdminRepositoryImpl implements AdminRepository {
     @Override
     public void blockUser(String userId) {
         User user = userRepository.findById(userId);
-        if (user != null) {
-            user.setIsBlocked(true);
-            userRepository.save(user);
+        if (user == null) {
+            throw new UserNotFoundException(userId);
         }
+        if (user.getIsBlocked()) {
+            throw new UserAlreadyBlockedException(userId);
+        }
+        user.setIsBlocked(true);
+        userRepository.save(user);
     }
 
     @Override
     public void deleteUser(String userId) {
+        if (userRepository.findById(userId) == null) {
+            throw new UserNotFoundException(userId);
+        }
         userRepository.delete(userId);
     }
 }
